pyproject.toml
[project]
name = "ipython-icat"
version = "0.2.3"
description = "IPython magic command for displaying images in kitty terminal"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ipython",
    "matplotlib",
    "pillow",
]
authors = [
    { name = "Tony Kabilan Okeke", email = "tonykabilanokeke@gmail.com" } 
]
keywords = ["ipython", "magic", "image", "display", "kitty", "icat"]

[project.urls]
Homepage = "https://github.com/Kabilan108/ipython-icat"

[tool.setuptools.packages.find]
where = ["."]
include = ["icat*"]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.1.1",
]
package = true

[project.scripts]
pycat = "icat.__init__:main"
